Refactor overlay detection into its own component separate from icon matching
